celpip_studio/requirements.txt
flask
pillow
numpy
scipy
pydub
edge-tts
kokoro-onnx
soundfile
huggingface_hub
fonttools
fpdf2
